Find GPU information from scontrol if not given from tres-per-node
Not sure if this is an update from slurm or specific to the VGG cluster
but now `squeue -o tres-per-node` flag seems to return something
that looks like `gres:gpu:<no_of_gpus>` in our version of slurm (21.08.07).

From looking at the existing code, it used to be something like
`gpu:<name-of-gpu>:<no-of-gpus>`.

Therefore this is updated so that the code notices if it is the newer
version of tres-per-node output. If so, gets the GPU information from
the detailed job info using `scontrol show job -dd <job_id>`.

I've tested this for jobs on our cluster and it seems to be working
but these area all using a single type of gpu on a single node
so it likely needs further changes for cases where people are
using several different types of gpu & nodes for a single job, by
further parsing the `JOB_GRES` output from `scontrol show`.